pyproject.toml
[project]
name = "weather-avg-temp"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = ["requests>=2.32.0", "python-dotenv>=1.0.0"]

[tool.pytest.ini_options]
pythonpath = ["src"]